helper function for csv header object added
